requirements.txt
Django==2.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0